Version 0.8.9
Version 0.8.9 prints a more clear error message to serial when the config files are not found. It also prints that error message to the TFT. I also spend some time formatting the code in the sketch as well as the .js and HTML files. Also removed some duplicate strcpy lines.