Version 0.8.18
Version 0.8.18 Added support for Deep Sleep. Use #define touchInterruptPin and "sleepenable": true + "sleeptimer": 10 in wificonfig.json to enable Deep Sleep and set the sleeptimer in minutes. Also added support for a buzzer (piezo speaker) use #define speakerPin to select the pin where the speaker is connected. For both: when the #defines are commented out, they are not used.

Some other changes: debug.h is removed. And wificonfig.json is modified to add the option to change wether sleep is enabled and to set the sleep timer.