Version 0.8.19
Version 0.8.19 Added "info" screen to the settings menu. Changed to including Javascript for the configurator in the index.html file itself. This makes the code less clear but reduces the number of requests to the ESP32. From my experience this results in less "hanging" and a bit faster loading of the configurator. Also added the spacebar as an option in the configurator.