Version 0.9.6
Version 0.9.6 Add latchting to different image instead of just a dot. This version also adds the ability to change the WiFi mode to be an Access Point and the ability to change the SSID and Password via the serial monitor. Furthermore, the touch calibration can be reset. See docs for usage!