Version 0.9.7
Bug fix: Changing colour causes reboot when entering config mode. "Combos"  dissapeared and are now back. Move the FreeTouchDeck logo to the "logos" folder so it can be deleted to create more space. Fixed the latching dot to fall within the boundaries of a button when using a 320 * 240 screen. Remember to re-upload the data folder beacuse HTML change. Configs from 0.9.6 can be used so maybe download them first!