Many changes!
support for old style config was dropped
Colors are now stored as RGB888 values internally and are only converted when needed. Originally, I was keeping the values as RGB565, which caused loss of data when converting back to RGB888 (e.g. how do you recognize white when shifting 5 bits back into an 8 bit value?).
Action parsing is now done in a new class of its own, ActionsSequences, which keeps the original text representation of a given action as well as individual internal parsed objects that are actually going to be executed
When drawing label buttons (e.g. no logo/icon specified), the system will see if the default font is too large and try to use a smaller font if possible.
This was made possible with the creation of a couple of new functions: InitFontsTable, SetFont, SetFont, SetDefaultFont, SetSmallerFont, SetSmallestFont
A bit of declutter on the logo handling bit. I'm still not super happy about the current state, but for now it is good enough
Action parsing is now solely using FREETEXT approach. Multiple action lines are possible under each button, and each line will automatically release control keys if required.
Below is an actual example taken from the updated "menus.json"
{
  "label":"Release",
  "backgroundcolor":"#66D2D6",
  "actions":["This is some text{TAB}After tab key{LEFT_SHIFT}some capitalized text",
  ". And this text has the shif key released automatically. {RETURN}",
  "{LEFT_SHIFT}some more capitalized text{RELEASEALL}. Followed by a manual release"]
}
Action types are now internally used only; no need to specify them anymore. They are now reduced to KEYBOARD or LOCAL and the parser takes care of recognizing the proper type when analyzing a token
There's a better tokenizer now; when parsing a command, tokens are automatically split into a list of strings. The first string (position 0) is, by default, the action name. Users wanting to experiment with their own actions could define callbacks that have an arbitrary number of parameters (memory permits it). Actions thus have a member called Parameters, and it is possible to access individual values via the following new members
const char *ActionName();
const char *FirstParameter();
const char *SecondParameter();
const char *ThirdParameter();
std::string& GetParameter(int index);
Parameters are stored as string rather than char *. It uses more memory, but is more user friendly.
For example, checking if we're doing brightness up or brightness down with a char * would be like this
if (strcmp(action->GetParameter(0), "BRIGHTNESS_UP") == 0)
whereas checking with string is more straightforward
if (action->GetParameter(0)== "BRIGHTNESS_UP")
In an attempt to see if picture rendering speed could be improved, I moved the image draw function to IRAM_ATTR. Not sure if it will help (after all, we're reading line by line from flash memory), but it's a start
All local actions (brightness, calling the info screen, etc.) are now defined as regular "UserActions", which means a ton of examples are there for anyone to leverage