add "memstatsdelay" configuration to reduce mem stats print to serial
Currently the memory stats are printed to the serial every 5s.  The
information is useful when debugging memory issues, but can get in
the way when debugging other issues.  This allows for a
"memstatusdelay" generalconfig option to change the frequency the
memory stats are printed, the value is in milliseconds (default value is
5000).  To disable the memory stats print entirely you can set the value
to "0"